pyproject.toml
[project]
name = "larisin-ai"
version = "0.1.0"
description = "Backend and AI service for Larisin AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.136.0",
    "uvicorn[standard]>=0.27.0",
    "openai>=2.30.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
    "email-validator>=2.1.0.post1",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "requests>=2.33.0",
    "azure-cosmos>=4.0.0"
]

[tool.uv.workspace]
members = [
    "larisin-pkg",
]
